third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-fallback-datalist-animations.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] select::picker(select) should support author provided top layer animations.
  assert_equals: All 4 of the CSS rules should have been parsed. expected 4 but got 1
Harness: the test ran to completion.
